Refactor: fold VerifySeedPhraseCoordinator into BackupSeedPhraseCoordinator.swift
